Corrected exercise indonesian_pronouns.md
There are two pronouns for the first person plural: kita (inclusive),
and kami (exclusive). Tina isn't going to Jakarta with Peter and Dita
(Tina is not included = use kami), neither is she returning with them
(Tina is not included = use kami). But, once Peter and Dita are back,
they go all together to a restaurant. In this case, Tina is included, so
use kita.